% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{aovMcomper2}
\alias{aovMcomper2}
\title{Perform ANOVA and run multiple comparisons and order in order to mean value of treatment}
\usage{
aovMcomper2(data = data_wt, i = 3, method_Mc = "Tukey")
}
\arguments{
\item{data}{a data.frame contain the input data}

\item{i}{col index wtich need to test}

\item{method_Mc}{which specify the method of multiple comparisons.included:"Tukey","LSD","SNK","Duncan","scheffe","scheffe".}
}
\value{
data frame
}
\description{
Perform ANOVA and run multiple comparisons and order in order to mean value of treatment
}
\examples{
# data(data_wt)
result= aovMcomper2 (data = data_wt, i= 5,method_Mc = "Tukey")
# utput result for multiple comparison results
result[[1]]
}
\references{
Yuan J, Zhao J, Wen T, Zhao M, Li R, Goossens P, Huang Q, Bai Y, Vivanco JM, Kowalchuk GA, Berendsen RL, Shen Q
Root exudates drive the soil-borne legacy of aboveground pathogen infection
Microbiome 2018,DOI: \url{doi: 10.1186/s40168-018-0537-x}
}
\author{
Contact: Tao Wen \email{2018203048@njau.edu.cn} Jun Yuan \email{junyuan@njau.edu.cn}
}
