% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BoxP.R
\name{aovMuiBoxP}
\alias{aovMuiBoxP}
\title{Variance test or non-parametric test results visualization, using boxplot}
\usage{
aovMuiBoxP(
  data = data_wt,
  i = 3,
  sig_show = "line",
  result = result,
  ns = FALSE
)
}
\arguments{
\item{data}{a data.frame contain the input data}

\item{i}{col index wtich need to test}

\item{sig_show}{Distinctive display, "abc" or "line"}

\item{result}{output from aovMcomper or KwWlx. You can also import result calculated from other software (a data frame)}

\item{ns}{Logical value, whether to display insignificant marks}
}
\value{
data frame
}
\description{
Variance test or non-parametric test results visualization, using boxplot
}
\examples{
# data(data_wt)
result = KwWlx(data = data_wt, i= 4)
PlotresultBox = aovMuiBoxP(data = data_wt, i= 4,sig_show ="abc",result = result[[1]])
# utput result
p = PlotresultBox[[1]]
p
}
\references{
Yuan J, Zhao J, Wen T, Zhao M, Li R, Goossens P, Huang Q, Bai Y, Vivanco JM, Kowalchuk GA, Berendsen RL, Shen Q
Root exudates drive the soil-borne legacy of aboveground pathogen infection
Microbiome 2018,DOI: \url{doi: 10.1186/s40168-018-0537-x}
}
\author{
Contact: Tao Wen \email{2018203048@njau.edu.cn} Jun Yuan \email{junyuan@njau.edu.cn}
}
