% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/es_anova.R
\name{es_anova}
\alias{es_anova}
\title{Easy anova}
\usage{
es_anova(data, y, ...)
}
\arguments{
\item{data}{Data to be used}

\item{y}{Dependent variable}

\item{...}{qualitative independent variable}
}
\description{
\code{es_anova()} returns a simple data frame containing the p-values calculated by anova
}
