% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/es_dataset.R
\name{es_dataset}
\alias{es_dataset}
\title{Formatting the data for Tukey or Dunnett}
\usage{
es_dataset(data, value, group)
}
\arguments{
\item{data}{Data to be used}

\item{value}{Dependent variable}

\item{group}{qualitative independent variable}
}
\description{
Extract the dependent and independent variables and order the factors
}
