% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/es_tuky_dunnett.R
\name{es_dunnett}
\alias{es_dunnett}
\title{Easy Dunnett test}
\usage{
es_dunnett(data, vx, fx)
}
\arguments{
\item{data}{Data to be used}

\item{vx}{Dependent variable}

\item{fx}{Independent variable}
}
\value{
Data frame with p-values in the Dunnett test
}
\description{
\code{es_dunnett} returns a simple data frame containing the p-values calculated by the dunnet test
}
