% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/es_dunnett.R
\name{es_dunnett_summary}
\alias{es_dunnett_summary}
\title{Summary of easy Dunnett's test}
\usage{
es_dunnett_summary(data, value, group)
}
\arguments{
\item{data}{Data to be used}

\item{value}{Dependent variable}

\item{group}{qualitative independent variable}
}
\value{
Data frame with p-values in the Dunnett's test
}
\description{
\code{es_dunnett_summary()} returns a simple data frame containing the p-values calculated by the Dunnett's test
}
