% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/es_t_test.R
\name{es_t_test}
\alias{es_t_test}
\title{Easy t-test}
\usage{
es_t_test(data, value, group, method = "welch")
}
\arguments{
\item{data}{Data to be used}

\item{value}{Dependent variable}

\item{group}{qualitative independent variable}

\item{method}{student or welch. Default is welch}
}
\value{
Data frame with p-values in the t-test
}
\description{
\code{es_t_test()} returns a simple data frame containing the p-values calculated by the t-test
}
