% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/es_tuky_dunnett.R
\name{es_tukey}
\alias{es_tukey}
\title{Easy tukey test}
\usage{
es_tukey(data, vx, fx, sig_level)
}
\arguments{
\item{data}{Data to be used}

\item{vx}{Dependent variable}

\item{fx}{Independent variable}

\item{sig_level}{Significance level}
}
\value{
Data frame containing the alphabet assigned by the tukey test
}
\description{
\code{es_tukey} returns a simple data frame containing the alphabet assigned
by the tukey test
}
