% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/es_tukey.R
\name{es_tukey_core}
\alias{es_tukey_core}
\title{The core of the Tukey's test}
\usage{
es_tukey_core(data, value, group, sig_level)
}
\arguments{
\item{data}{Data to be used}

\item{value}{Dependent variable}

\item{group}{qualitative independent variable}

\item{sig_level}{Significance level}
}
\description{
The core of the Tukey's test
}
