% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/es_tukey.R
\name{es_tukey_summary}
\alias{es_tukey_summary}
\title{Summary of easy Tukey's test}
\usage{
es_tukey_summary(data, value, group, sig_level)
}
\arguments{
\item{data}{Data to be used}

\item{value}{Dependent variable}

\item{group}{qualitative independent variable}

\item{sig_level}{Significance level}
}
\value{
Data frame containing the alphabet assigned by the Tukey's test
}
\description{
\code{es_tukey_summary()} returns a simple data frame containing the alphabet assigned by the Tukey's test
}
