% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FacetMuiPlotresultBar.R
\name{FacetMuiPlotresultBar}
\alias{FacetMuiPlotresultBar}
\title{Use facets to show the results of variance analysis, bar plot selected}
\usage{
FacetMuiPlotresultBar(
  data = data_wt,
  num = c(4:6),
  result = result,
  sig_show = "abc",
  ncol = 3,
  fac.level = NULL
)
}
\arguments{
\item{data}{a data.frame contain the input data;Enter the data frame, the first column is the sample number, the second column is the group, note that the group label must be set to group, and the third column and later are the indicators for measurement or collection.}

\item{num}{col index wtich need to test}

\item{result}{output from aovMcomper or KwWlx. You can also import result calculated from other software (a data frame)}

\item{sig_show}{Distinctive display, "abc" or "line"}

\item{ncol}{If you choose faceted display, choose to place several graphics per line}
}
\value{
list
}
\description{
Use facets to show the results of variance analysis, bar plot selected
}
\examples{
# data(data_wt)
result = MuiKwWlx(data = data_wt,num = c(4:6))
result1 = FacetMuiPlotresultBar(data = data_wt,num = c(4:6),result = result,sig_show ="abc",ncol = 2 )
result1[[1]]
}
\references{
Yuan J, Zhao J, Wen T, Zhao M, Li R, Goossens P, Huang Q, Bai Y, Vivanco JM, Kowalchuk GA, Berendsen RL, Shen Q
Root exudates drive the soil-borne legacy of aboveground pathogen infection
Microbiome 2018,DOI: \url{doi: 10.1186/s40168-018-0537-x}
}
\author{
Contact: Tao Wen \email{2018203048@njau.edu.cn} Jun Yuan \email{junyuan@njau.edu.cn}
}
