% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GroupBar.R
\name{GroupBar}
\alias{GroupBar}
\title{Variance test or non-parametric test results visualization, using boxbar}
\usage{
GroupBar(
  data = data_wt,
  num = c(4:6),
  result = result,
  sig_show = "abc",
  ylimit = 1.2
)
}
\arguments{
\item{data}{a data.frame contain the input data}

\item{num}{index col which need plot}

\item{result}{output from aovMcomper or KwWlx. You can also import result calculated from other software (a data frame)}

\item{sig_show}{Distinctive display, "abc" or we could find in the future}

\item{ylimit}{Adjust the y-axis coordinate range}
}
\value{
list
}
\description{
Variance test or non-parametric test results visualization, using boxbar
}
\examples{
# data(data_wt)
result = MuiKwWlx(data = data_wt,num = c(4:6))
result = GroupBar(data = data_wt,num = c(4:6),result = result,sig_show ="abc")
# output result
result[[1]]

}
\references{
Yuan J, Zhao J, Wen T, Zhao M, Li R, Goossens P, Huang Q, Bai Y, Vivanco JM, Kowalchuk GA, Berendsen RL, Shen Q
Root exudates drive the soil-borne legacy of aboveground pathogen infection
Microbiome 2018,DOI: \url{doi: 10.1186/s40168-018-0537-x}
}
\author{
Contact: Tao Wen \email{2018203048@njau.edu.cn} Jun Yuan \email{junyuan@njau.edu.cn}
}
