% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KwWlxPair.R
\name{KwWlxPair}
\alias{KwWlxPair}
\title{Easy for Non-parametric test}
\usage{
KwWlxPair(data = data_wt, i = 3, method = "wilcox.test", paired = FALSE)
}
\arguments{
\item{data}{a data.frame contain the input data}

\item{i}{col index wtich need to test}

\item{method}{"t.test" or "wilcox.test" could be selected by method}

\item{paired}{a logical indicating whether you want a paired test. Used in t.test and in wilcox.test.}
}
\value{
data frame
}
\description{
Easy for Non-parametric test
}
\examples{
# data(data_wt)
result = KwWlx2(data = data_wt, i= 4,method = "t.test",paired = FALSE)
result[[1]]
}
\references{
Yuan J, Zhao J, Wen T, Zhao M, Li R, Goossens P, Huang Q, Bai Y, Vivanco JM, Kowalchuk GA, Berendsen RL, Shen Q
Root exudates drive the soil-borne legacy of aboveground pathogen infection
Microbiome 2018,DOI: \url{doi: 10.1186/s40168-018-0537-x}
}
\author{
Contact: Tao Wen \email{2018203048@njau.edu.cn} Jun Yuan \email{2018203048@njau.edu.cn}
}
