% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MuiaovMcomper.R
\name{MuiaovMcomper}
\alias{MuiaovMcomper}
\title{Perform ANOVA and run multiple comparisons for more sets of data}
\usage{
MuiaovMcomper(data = data_wt, num = c(4:6), method_Mc = "Tukey")
}
\arguments{
\item{data}{a data.frame contain the input data}

\item{num}{index col which to test normality and homogeneity}

\item{method_Mc}{which specify the method of multiple comparisons.included:"Tukey","LSD","SNK","Duncan","scheffe","scheffe".}
}
\value{
data frame
}
\description{
Perform ANOVA and run multiple comparisons for more sets of data
}
\examples{
# data(data_wt)
result = MuiaovMcomper(data = data_wt,num = c(4:6),method_Mc = "Tukey")
# utput result for multiple comparison results
result
}
\references{
Yuan J, Zhao J, Wen T, Zhao M, Li R, Goossens P, Huang Q, Bai Y, Vivanco JM, Kowalchuk GA, Berendsen RL, Shen Q
Root exudates drive the soil-borne legacy of aboveground pathogen infection
Microbiome 2018,DOI: \url{doi: 10.1186/s40168-018-0537-x}
}
\author{
Contact: Tao Wen \email{2018203048@njau.edu.cn} Jun Yuan \email{2018203048@njau.edu.cn}
}
