% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MuiHeatmapBubplot.R
\name{MuiHeatmapBubplot}
\alias{MuiHeatmapBubplot}
\title{MuiHeatmapBubplot}
\usage{
MuiHeatmapBubplot(
  data = data,
  i,
  result = result,
  sample = FALSE,
  scale = TRUE,
  col_cluster = TRUE,
  row_cluster = TRUE,
  label = TRUE
)
}
\arguments{
\item{data}{a data.frame contain the input data}

\item{i}{col index wtich need to test}

\item{result}{output from aovMcomper or KwWlx. You can also import result calculated from other software (a data frame)}

\item{sample}{sample or group}

\item{col_cluster}{only used for sample = TRUE}

\item{row_cluster}{only used for sample = TRUE}

\item{label}{only used for sample = TRUE}

\item{errbar}{add a distinctive label, TRUE or FEASE could be selected}
}
\value{
list
}
\description{
MuiHeatmapBubplot
}
\examples{
# data(data_wt)
result = MuiKwWlx(data = data_wt,num = c(4,9,8))
res <- MuiHeatmapBubplot(data = data_wt,i = c(4,9,8) ,result = result,
                     sample  = TRUE)
res[[1]]
}
\references{
Yuan J, Zhao J, Wen T, Zhao M, Li R, Goossens P, Huang Q, Bai Y, Vivanco JM, Kowalchuk GA, Berendsen RL, Shen Q
Root exudates drive the soil-borne legacy of aboveground pathogen infection
Microbiome 2018,DOI: \url{doi: 10.1186/s40168-018-0537-x}
}
\author{
Contact: Tao Wen \email{2018203048@njau.edu.cn} Jun Yuan \email{junyuan@njau.edu.cn} PengHao Xie \email{2019103106@njau.edu.cn}
}
