% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{MuiKwWlx2}
\alias{MuiKwWlx2}
\title{Multi-group data for non-parametric test}
\usage{
MuiKwWlx2(data = data_wt, num = c(4:6), p.adjust.method = "none")
}
\arguments{
\item{data}{a data.frame contain the input data and order in order to mean value of treatment}

\item{num}{index col which need plot}
}
\value{
data frame
}
\description{
Multi-group data for non-parametric test
}
\examples{
# data(data_wt)
 result = MuiKwWlx(data = data_wt,num = c(4:6))
result
}
\references{
Yuan J, Zhao J, Wen T, Zhao M, Li R, Goossens P, Huang Q, Bai Y, Vivanco JM, Kowalchuk GA, Berendsen RL, Shen Q
Root exudates drive the soil-borne legacy of aboveground pathogen infection
Microbiome 2018,DOI: \url{doi: 10.1186/s40168-018-0537-x}
}
\author{
Contact: Tao Wen \email{2018203048@njau.edu.cn} Jun Yuan \email{junyuan@njau.edu.cn}
}
