% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MuiNorCV.R
\name{MuiNorCV}
\alias{MuiNorCV}
\title{Test normality and homogeneity of variance of inputmore sets of data}
\usage{
MuiNorCV(
  data = data_wt,
  num = c(4:6),
  method_cv = "leveneTest",
  normality_adjust = c("none", "bonferroni")
)
}
\arguments{
\item{data}{a data.frame contain the input data}

\item{num}{index col which to test normality and homogeneity}

\item{method_cv}{which method to test  variance, included "leveneTest","bartlett.test".}
}
\value{
data frame
}
\description{
Test normality and homogeneity of variance of inputmore sets of data
}
\examples{
# data(data_wt)
norCv = MuiNorCV(data = data_wt,num = c(4,5,6),method_cv = "leveneTest")
norCv
}
\references{
Yuan J, Zhao J, Wen T, Zhao M, Li R, Goossens P, Huang Q, Bai Y, Vivanco JM, Kowalchuk GA, Berendsen RL, Shen Q
Root exudates drive the soil-borne legacy of aboveground pathogen infection
Microbiome 2018,DOI: \url{doi: 10.1186/s40168-018-0537-x}
}
\author{
Contact: Tao Wen \email{2018203048@njau.edu.cn} Jun Yuan \email{junyuan@njau.edu.cn}
}
