% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MuiPlotresultBar.R
\name{MuiPlotresultBar}
\alias{MuiPlotresultBar}
\title{Visualize multiple sets of data  barplot}
\usage{
MuiPlotresultBar(
  data = data_wt,
  num = c(4:6),
  result = result,
  sig_show = "abc",
  path = "./Muibar/"
)
}
\arguments{
\item{data}{a data.frame contain the input data}

\item{num}{col index wtich need to test}

\item{result}{output from aovMcomper or KwWlx. You can also import result calculated from other software (a data frame)}

\item{sig_show}{Distinctive display, "abc" or "line"}

\item{path}{creat a folder to save the single plot.}
}
\value{
list
}
\description{
Visualize multiple sets of data  barplot
}
\examples{
# data(data_wt)
result = MuiKwWlx(data = data_wt,num = c(4:6))
MuiPlotresultBar(data = data_wt,num = c(4:6),result = result ,sig_show ="line")

}
\references{
Yuan J, Zhao J, Wen T, Zhao M, Li R, Goossens P, Huang Q, Bai Y, Vivanco JM, Kowalchuk GA, Berendsen RL, Shen Q
Root exudates drive the soil-borne legacy of aboveground pathogen infection
Microbiome 2018,DOI: \url{doi: 10.1186/s40168-018-0537-x}
}
\author{
Contact: Tao Wen \email{2018203048@njau.edu.cn} Jun Yuan \email{junyuan@njau.edu.cn}
}
