% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MuiPlotresultBox.R, R/MuiPlotresultBox2.R
\name{MuiPlotresultBox}
\alias{MuiPlotresultBox}
\title{Visualize multiple sets of data Scatter plot + boxplot}
\usage{
MuiPlotresultBox(
  data = data_wt,
  num = c(4:6),
  result = result,
  sig_show = "abc",
  path = "./Muibox/"
)

MuiPlotresultBox(
  data = data_wt,
  num = c(4:6),
  result = result,
  sig_show = "abc",
  path = "./Muibox/"
)
}
\arguments{
\item{data}{a data.frame contain the input data}

\item{num}{col index wtich need to test}

\item{result}{output from aovMcomper or KwWlx. You can also import result calculated from other software (a data frame)}

\item{sig_show}{Distinctive display, "abc" or "line"}

\item{path}{creat a folder to save the single plot.}
}
\value{
list

list
}
\description{
Visualize multiple sets of data Scatter plot + boxplot

Visualize multiple sets of data Scatter plot + boxplot
}
\examples{
# data(data_wt)
result = KwWlx(data = data_wt, i= 4)
MuiPlotresultBox(data = data_wt,num = c(4:6),result = result,sig_show ="abc")

# data(data_wt)
result = KwWlx(data = data_wt, i= 4)
MuiPlotresultBox(data = data_wt,num = c(4:6),result = result,sig_show ="abc")

}
\references{
Yuan J, Zhao J, Wen T, Zhao M, Li R, Goossens P, Huang Q, Bai Y, Vivanco JM, Kowalchuk GA, Berendsen RL, Shen Q
Root exudates drive the soil-borne legacy of aboveground pathogen infection
Microbiome 2018,DOI: \url{doi: 10.1186/s40168-018-0537-x}

Yuan J, Zhao J, Wen T, Zhao M, Li R, Goossens P, Huang Q, Bai Y, Vivanco JM, Kowalchuk GA, Berendsen RL, Shen Q
Root exudates drive the soil-borne legacy of aboveground pathogen infection
Microbiome 2018,DOI: \url{doi: 10.1186/s40168-018-0537-x}
}
\author{
Contact: Tao Wen \email{2018203048@njau.edu.cn} Jun Yuan \email{junyuan@njau.edu.cn}

Contact: Tao Wen \email{2018203048@njau.edu.cn} Jun Yuan \email{junyuan@njau.edu.cn}
}
