% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MuiPlotStackBar.R
\name{MuiPlotStackBar}
\alias{MuiPlotStackBar}
\title{Stacked histograms add error bars, marked by significance}
\usage{
MuiPlotStackBar(
  data = data,
  i,
  result = result,
  errbar = TRUE,
  add_abc = TRUE,
  order = FALSE,
  sample = FALSE,
  ...
)
}
\arguments{
\item{data}{a data.frame contain the input data}

\item{i}{col index wtich need to test}

\item{result}{output from aovMcomper or KwWlx. You can also import result calculated from other software (a data frame)}

\item{errbar}{add a distinctive label, TRUE or FEASE could be selected}

\item{add_abc}{add sig abc T or F}

\item{order}{order with colnames according to value}

\item{sample}{T or F show sample data not group data}
}
\value{
list
}
\description{
Stacked histograms add error bars, marked by significance
}
\examples{
# data(data_wt)
result = MuiKwWlx(data = data_wt,num = c(4,9,8))
res <- MuiPlotStackBar(data = data,i = c(4,9,8) ,result = result)
# utput result
res[[1]]
}
\references{
Yuan J, Zhao J, Wen T, Zhao M, Li R, Goossens P, Huang Q, Bai Y, Vivanco JM, Kowalchuk GA, Berendsen RL, Shen Q
Root exudates drive the soil-borne legacy of aboveground pathogen infection
Microbiome 2018,DOI: \url{doi: 10.1186/s40168-018-0537-x}
}
\author{
Contact: Tao Wen \email{2018203048@njau.edu.cn} Jun Yuan \email{junyuan@njau.edu.cn} PengHao Xie \email{2019103106@njau.edu.cn}
}
