% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_methods.R
\name{summary.easystat_result}
\alias{summary.easystat_result}
\title{Summarize an EasyStat Result Object}
\usage{
\method{summary}{easystat_result}(object, ...)
}
\arguments{
\item{object}{An \code{"easystat_result"} object.}

\item{...}{Passed to \code{print.easystat_result}.}
}
\description{
Summarize an EasyStat Result Object
}
