% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easystat_theme.R
\name{theme_easystat}
\alias{theme_easystat}
\title{Apply EasyStat ggplot2 Theme}
\usage{
theme_easystat(base_size = 12, legend_position = "right")
}
\arguments{
\item{base_size}{Base font size. Default \code{12}.}

\item{legend_position}{Where to place the legend. Default \code{"right"}.}
}
\value{
A \code{ggplot2::theme} object.
}
\description{
Adds a clean, professional EasyStat visual theme to any ggplot2 object.
}
