% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WelchAov.R
\name{WelchAov}
\alias{WelchAov}
\title{Perform Welch's ANOVA and Games-Howell post-hoc test for heterogeneous variance data}
\usage{
WelchAov(data = data_wt, i = 3)
}
\arguments{
\item{data}{a data.frame; first column is sample ID, second column is group,
subsequent columns are response variables.}

\item{i}{col index or column name of the variable to test.}
}
\value{
list: [[1]] data.frame of CLD letters with group column;
              [[2]] Welch's ANOVA summary;
              [[3]] Games-Howell pairwise results
}
\description{
When data passes normality test but fails homogeneity of variance test,
Welch's ANOVA is more appropriate than standard one-way ANOVA.
Post-hoc comparisons use Games-Howell test, which does not assume equal variances.
}
\examples{
# data(data_wt)
result = WelchAov(data = data_wt, i = 5)
result[[1]]
}
\author{
Contact: Tao Wen \email{2018203048@njau.edu.cn}
}
