% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easystat-package.R
\docType{package}
\name{easystat-package}
\alias{easystat}
\alias{easystat-package}
\title{easystat: Easy to perform statistical tests}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Yusuke Fukuda \email{tonegawa07@gmail.com}

}
\keyword{internal}
