% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/es_tuky_dunnett.R
\name{es_dataset}
\alias{es_dataset}
\title{Formatting the dataset & setting the factor}
\usage{
es_dataset(data, vx, fx)
}
\arguments{
\item{data}{Data to be used}

\item{vx}{Dependent variable}

\item{fx}{Independent variable}
}
\description{
Extract the dependent and independent variables and order the factors
}
