% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MuiStat.R
\name{MuiStat}
\alias{MuiStat}
\title{A complete single-factor difference analysis process for mui col data, including normal test, homogeneity analysis of variance, selection of variance analysis or non-parametric test, selection of candidate d visualization method.}
\usage{
MuiStat(
  data = data_wt,
  num = c(4, 5, 6),
  method_cv = "leveneTest",
  method_Mc = "Tukey",
  sig_show = "abc",
  ncol = 2,
  plot = "bar",
  plottype = "mui",
  path = "./output"
)
}
\arguments{
\item{data}{a data.frame contain the input data;Enter the data frame, the first column is the sample number, the second column is the group, note that the group label must be set to group, and the third column and later are the indicators for measurement or collection.}

\item{num}{col index wtich need to test}

\item{method_cv}{which method to test  variance, included "leveneTest","bartlett.test"}

\item{method_Mc}{which specify the method of multiple comparisons.included:"Tukey","LSD","SNK","Duncan","scheffe","scheffe".}

\item{sig_show}{Distinctive display, "abc" or "line"}

\item{ncol}{If you choose faceted display, choose to place several graphics per line}

\item{plot}{chose the plot type shuld be a barplot or boxplot or bar+box plot}

\item{plottype}{select a plot type, included "single" and "mui"}

\item{path}{creat a folder to save the single plot, if you seleted the model "single"}
}
\value{
list
}
\description{
A complete single-factor difference analysis process for mui col data, including normal test, homogeneity analysis of variance, selection of variance analysis or non-parametric test, selection of candidate d visualization method.
}
\examples{
# data(data_wt)
result = MuiStat(data = data_wt,num = c(4,5,6),method_cv = "leveneTest",method_Mc = "Tukey",sig_show  = "abc",ncol = 2,plot = "boxbar",plottype = "mui")
result[[1]]
}
\references{
Yuan J, Zhao J, Wen T, Zhao M, Li R, Goossens P, Huang Q, Bai Y, Vivanco JM, Kowalchuk GA, Berendsen RL, Shen Q
Root exudates drive the soil-borne legacy of aboveground pathogen infection
Microbiome 2018,DOI: \url{doi: 10.1186/s40168-018-0537-x}
}
\author{
Contact: Tao Wen \email{2018203048@njau.edu.cn} Jun Yuan \email{junyuan@njau.edu.cn}
}
