% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NorNorCVTest.R
\name{NorNorCVTest}
\alias{NorNorCVTest}
\title{Test normality and homogeneity of variance of Input data}
\usage{
NorNorCVTest(data = data_wt, i = 4, method_cv = "leveneTest", ...)
}
\arguments{
\item{data}{a data.frame contain the input data}

\item{i}{col index wtich need to test}

\item{method_cv}{which method to test  variance, included "leveneTest","bartlett.test".}
}
\value{
data frame
}
\description{
Test normality and homogeneity of variance of Input data
}
\examples{
# data(data_wt)
NorNorCVTest(data = data_wt, i= 4)
}
\references{
Yuan J, Zhao J, Wen T, Zhao M, Li R, Goossens P, Huang Q, Bai Y, Vivanco JM, Kowalchuk GA, Berendsen RL, Shen Q
Root exudates drive the soil-borne legacy of aboveground pathogen infection
Microbiome 2018,DOI: \url{doi: 10.1186/s40168-018-0537-x}
}
\author{
Contact: Tao Wen \email{2018203048@njau.edu.cn} Jun Yuan \email{junyuan@njau.edu.cn}
}
