% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ord_sig.R
\name{ord_sig}
\alias{ord_sig}
\title{Reverse the significance level of letter marks}
\usage{
ord_sig(data = da, ID = "groups")
}
\arguments{
\item{data}{a data.frame contain the input data}

\item{ID}{col index wtich need to reverse significance level,if NULL was marked, would transform to each col.}
}
\value{
data.frame
}
\description{
Reverse the significance level of letter marks
}
\examples{
# data(data_wt)
result = KwWlx(data = data_wt, i= 4)
datord <- ord_sig(data = result[[1]],ID = NULL)
}
\references{
Yuan J, Zhao J, Wen T, Zhao M, Li R, Goossens P, Huang Q, Bai Y, Vivanco JM, Kowalchuk GA, Berendsen RL, Shen Q
Root exudates drive the soil-borne legacy of aboveground pathogen infection
Microbiome 2018,DOI: \url{doi: 10.1186/s40168-018-0537-x}
}
\author{
Contact: Tao Wen \email{2018203048@njau.edu.cn} Jun Yuan \email{junyuan@njau.edu.cn}
}
