% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleStat.R
\name{SingleStat}
\alias{SingleStat}
\title{A complete single-factor difference analysis process for single col data, including normal test, homogeneity analysis of variance, selection of variance analysis or non-parametric test, selection of candidate d visualization method.}
\usage{
SingleStat(
  data = data_wt,
  plot = "bar",
  method_Mc = "Tukey",
  i = 4,
  sig_show = "abc"
)
}
\arguments{
\item{data}{a data.frame contain the input data.Enter the data frame, the first column is the sample number,
the second column is the group, note that the group label must be set to group, and the third column and later are the
indicators for measurement or collection.}

\item{method_Mc}{which specify the method of multiple comparisons.included:"Tukey","LSD","SNK","Duncan","scheffe","scheffe".}

\item{sig_show}{Distinctive display, "abc" or "line"}

\item{num}{col index wtich need to test}

\item{method_cv}{which method to test  variance, included "leveneTest","bartlett.test"}
}
\value{
list
}
\description{
A complete single-factor difference analysis process for single col data, including normal test, homogeneity analysis of variance, selection of variance analysis or non-parametric test, selection of candidate d visualization method.
}
\examples{
# data(data_wt)
result = SingleStat(data = data_wt,plot = "bar",method_Mc = "Tukey",i= 4,sig_show ="abc")
result[[1]]
}
\references{
Yuan J, Zhao J, Wen T, Zhao M, Li R, Goossens P, Huang Q, Bai Y, Vivanco JM, Kowalchuk GA, Berendsen RL, Shen Q
Root exudates drive the soil-borne legacy of aboveground pathogen infection
Microbiome 2018,DOI: \url{doi: 10.1186/s40168-018-0537-x}
}
\author{
Contact: Tao Wen \email{2018203048@njau.edu.cn} Jun Yuan \email{junyuan@njau.edu.cn}
}
